split dependecy sources, add composite source
