+++ src/dishka/dependency_source/__init__.py
@@ -1,14 +1,17 @@
 __all__ = [
     "DependencySource",
     "alias", "Alias",
-    "context_var", "from_context",
+    "context_var", "ContextVariable", "from_context",
     "decorate", "Decorator",
     "provide", "Factory", "FactoryType",
 ]
 
-from .alias import Alias, alias
-from .context_var import ContextVariable, from_context
-from .decorator import Decorator, decorate
-from .factory import Factory, FactoryType, provide
-
-DependencySource = Alias | Factory | Decorator | ContextVariable
+from .alias import Alias
+from .composite import DependencySource
+from .context_var import ContextVariable
+from .decorator import Decorator
+from .factory import Factory, FactoryType
+from .make_alias import alias
+from .make_context_var import from_context
+from .make_decorator import decorate
+from .make_factory import provide
