+++ src/dishka/dependency_source/decorator.py
@@ -1,15 +1,7 @@
-from collections.abc import (
-    Callable,
-)
-from typing import (
-    Any,
-    overload,
-)
-
 from dishka.entities.component import Component
 from dishka.entities.key import DependencyKey
 from dishka.entities.scope import BaseScope
-from .factory import Factory, make_factory
+from .factory import Factory
 
 
 class Decorator:
@@ -45,37 +37,3 @@ class Decorator:
 
     def __get__(self, instance, owner):
         return Decorator(self.factory.__get__(instance, owner))
-
-
-@overload
-def decorate(
-        *,
-        provides: Any = None,
-) -> Callable[[Callable], Decorator]:
-    ...
-
-
-@overload
-def decorate(
-        source: Callable | type,
-        *,
-        provides: Any = None,
-) -> Decorator:
-    ...
-
-
-def decorate(
-        source: Callable | type | None = None,
-        provides: Any = None,
-) -> Decorator | Callable[[Callable], Decorator]:
-    if source is not None:
-        return Decorator(make_factory(
-            provides=provides, scope=None, source=source, cache=False,
-        ))
-
-    def scoped(func):
-        return Decorator(make_factory(
-            provides=provides, scope=None, source=func, cache=False,
-        ))
-
-    return scoped
